refactor: extract order form logic
