+++ client/src/components/Order/FormModal.jsx
@@ -19,77 +19,26 @@ import {
   NumberInputStepper,
   Select,
   Textarea,
-  useDisclosure,
-  useToast,
 } from '@chakra-ui/react';
+import useOrderForm from 'hooks/useOrderForm.jsx';
 import PropTypes from 'prop-types';
-import React, { Fragment, useRef, useState } from 'react';
-import { useQuery } from 'react-query';
-import { getMaterialList, getOrders, postOrders } from 'utils/api/orders.js';
-
-// generates a random int within ranges
-const getRandomInt = (min, max) => {
-  return Math.floor(Math.random() * (max - min)) + min;
-};
+import React, { Fragment } from 'react';
 
 const FormModal = ({ showButton = false }) => {
   // TODO: use the endpoint to get this constant list of materials rather than hard coding
   const materialTypes = ['rubber', 'aluminum', 'steel', 'copper', 'plastic', 'leather'];
-  const manufacturer = useRef('');
-  const location = useRef('');
-  const note = useRef('');
-  const toast = useToast();
-  const [material, setMaterial] = useState(null);
-  const [cost, setCost] = useState(0);
-  const materialCost = useQuery('orders/materialList', getMaterialList);
-  const { refetch } = useQuery('orders', getOrders);
-  const { isOpen, onOpen, onClose } = useDisclosure();
-  const quantityRef = useRef(1);
-
-  // sets material and cost states
-  const handleMaterial = (e) => {
-    const choice = e.target.value;
-    setMaterial(choice);
-    setCost(materialCost.data.data[choice].cost);
-  };
-
-  const handleSubmit = async () => {
-    const orderTime = new Date();
-    const randomDeliveryDate = new Date().getDate() + getRandomInt(1, 8); // random delivery date within a week
-
-    try {
-      await postOrders({
-        materialType: material,
-        cost: cost,
-        quantity: quantityRef.current.value,
-        deliveryDate: randomDeliveryDate,
-        orderDate: orderTime,
-        manufacturerName: manufacturer.current.value,
-        vendorLocation: location.current.value,
-        status: 'Pending',
-        note: note.current.value,
-      });
-      toast({
-        title: 'Order Placed',
-        description: 'The order has been placed',
-        status: 'success',
-        duration: 9000,
-        isClosable: true,
-      });
-    } catch {
-      toast({
-        position: 'top',
-        title: 'An error occurred.',
-        description: 'Unable to place order.',
-        status: 'error',
-        duration: 2000,
-        isClosable: true,
-      });
-    }
-    refetch();
-    onClose();
-    setMaterial(null);
-  };
+  const {
+    handleMaterial,
+    handleSubmit,
+    manufacturer,
+    location,
+    note,
+    quantityRef,
+    isOpen,
+    onOpen,
+    onClose,
+    material,
+  } = useOrderForm();
 
   return (
     <>
